add jsdom env for testing and finish tests
