--- source/jest.config.js
@@ -1,6 +1,0 @@
-export default {
-    transform: {
-        "^.+\\.[t|j]sx?$": "babel-jest"
-    },
-    verbose: true
-};
