+++ source/package.json
@@ -20,9 +20,10 @@
     "escape-string-regexp": "^5.0.0",
     "eslint": "^9.2.0",
     "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "jest-puppeteer": "^10.0.1",
-    "puppeteer": "^22.10.0",
-    "jsdoc": "^4.0.3"
+    "jsdoc": "^4.0.3",
+    "puppeteer": "^22.10.0"
   },
   "type": "module"
 }
